improve codes.

+++ src/LNLib/Mathematics/Utils/MathUtils.cpp
@@ -92,7 +92,7 @@ int LNLib::MathUtils::Factorial(int number)
 
 double LNLib::MathUtils::Binomial(int number, int i)
 {
-    return Factorial(number) / (Factorial(i) * Factorial(number - i));
+    return static_cast<double>(Factorial(number)) / (static_cast<double>(Factorial(i)) * Factorial(number - i));
 }
 
 double LNLib::MathUtils::ComputerCubicEquationsWithOneVariable(double cubic, double quadratic, double linear, double constant)
